requirements.txt
asgiref==3.8.1
Django==5.1.3
django-crispy-forms==2.3
sqlparse==0.5.2
tzdata==2024.2

Django>=5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0